Add files via upload
Version 0.1

The player is able to move an collect food to raise the score. The playground is being filled with lava, but at this state at a bad timing and placement.

The lave is able to spawn on and in the near of the player, which needs to be fixed. Also the lava spawn-timing is to highat the moment and needs to be set at about 20 seconds. Beside, the function of lava expanding is not given.

The game has no menu, option, game over / replay screen.
It should also be able to save the score in a highscore array.